records-ui: Added forms, views, templates for contact_the_owner functionality
