+++ zenodo/modules/pages/templates/zenodo_pages/contact_the_owner.html
@@ -1,0 +1,66 @@
+{#
+# This file is part of Zenodo.
+# Copyright (C) 2017 CERN.
+#
+# Zenodo is free software; you can redistribute it
+# and/or modify it under the terms of the GNU General Public License as
+# published by the Free Software Foundation; either version 2 of the
+# License, or (at your option) any later version.
+#
+# Zenodo is distributed in the hope that it will be
+# useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with Zenodo; if not, write to the
+# Free Software Foundation, Inc., 59 Temple Place, Suite 330, Boston,
+# MA 02111-1307, USA.
+#
+# In applying this license, CERN does not
+# waive the privileges and immunities granted to it by virtue of its status
+# as an Intergovernmental Organization or submit itself to any jurisdiction.
+-#}
+
+{% extends "zenodo_theme/page.html" %}
+
+{% from "invenio_userprofiles/settings/_macros.html" import render_field %}
+
+{%- block page_body %}
+<div class="container">
+  <div class="row">
+    <div class="col-md-3"></div>
+    <div class="col-md-6">
+      <h2>Get help with Zenodo</h2>
+      <hr />
+      <div class="panel-body">
+      <form action="" method="POST" enctype="multipart/form-data">
+        {% if not current_user.is_authenticated %}
+          It is recommended to login first.
+        {% endif %}
+
+        {{ form.csrf_token }}
+        {{ render_field(form.name) }}
+
+        {% if current_user.is_authenticated %}
+          Zenodo staff will get back to you by your Zenodo email: <b>{{ form.email.data }}</b>
+        {% else %}
+          {{ render_field(form.email) }}
+        {% endif %}
+
+        {{ render_field(form.message) }}
+        
+        <br />
+
+        {{ render_field(form.recaptcha) }}
+
+        <button class="btn btn-block btn-success" type="submit">
+          Send Request
+        </button>
+      </form>
+      </div>
+    </div>
+    <div class="col-md-3"></div>
+  </div>
+</div>
+{%- endblock page_body %}
